dbbuildig metric

+++ src/player/Player.h
@@ -21,7 +21,7 @@ public:
 
 	Resources& getResources() { return resources; }
 	Possession& getPossession() { return possession; }
-	int getNation() const { return dbNation->id; }
+	short getNation() const { return dbNation->id; }
 	char getTeam() const { return team; }
 	char getId() const { return id; }//TODO bug id playera a jego index to czêsto nie to samo
 	int getColor() const { return color; }
